LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigation-traverseTo-navigates-top-and-same-doc-child-and-cross-doc-child-expected.txt



FAIL navigation.traverseTo() can navigate 3 frames of different types with correct navigate event cancelable values assert_equals: expected 2 but got 1
